Arts 265 (#13)
* [ARTS-265]
Removing security config for the service as this will be an internal call and any security will be external to these files.

* [ARTS-265]
Removing security config for the service as this will be an internal call and any security will be external to these files. Fixing merge conflicts

* [ARTS-265]
Removing security config for the service as this will be an internal call and any security will be external to these files. Fixing merge conflicts

* [ARTS-265]
Removing security config for the service as this will be an internal call and any security will be external to these files. Fixing merge conflicts. Readding swagger code from webconfig - missed during the auth purge.